backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
firebase-admin==6.5.0
requests==2.31.0
python-dotenv==1.0.1
python-multipart==0.0.9
